requirements.txt
flask==2.3.3
flask-cors==4.0.0
paho-mqtt==1.6.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
gunicorn==21.2.0
python-dotenv==1.0.0
pandas==2.0.3
grafana-api==1.0.3
flask-socketio==5.3.4
eventlet==0.33.3